Move label types to folder
