+++ packages/x-charts/src/BarChart/BarLabel/types.tsx
@@ -1,0 +1,48 @@
+import { SeriesId } from '../../models/seriesType/common';
+import type { BarLabelClasses } from './barLabelClasses';
+
+export interface BarLabelOwnerState {
+  seriesId: SeriesId;
+  dataIndex: number;
+  color: string;
+  isFaded: boolean;
+  isHighlighted: boolean;
+  classes?: Partial<BarLabelClasses>;
+}
+
+export type BarLabelRootProps = Omit<React.ComponentPropsWithoutRef<'text'>, 'id'> & {
+  ownerState: BarLabelOwnerState;
+};
+
+export type BarItem = {
+  /**
+   * The series id of the bar.
+   */
+  seriesId: SeriesId;
+  /**
+   * The index of the data point in the series.
+   */
+  dataIndex: number;
+  /**
+   * The value of the data point.
+   */
+  value: number | null;
+};
+
+export type BarLabelContext = {
+  bar: {
+    /**
+     * The height of the bar. Useful if you want to show the label only when the bar is big enough.
+     */
+    height: number;
+    /**
+     * The width of the bar. Useful if you want to show the label only when the bar is big enough.
+     */
+    width: number;
+  };
+};
+
+export type BarLabelFunction = (
+  item: BarItem,
+  context: BarLabelContext,
+) => string | null | undefined;
